requirements.txt

cassandra-driver==3.24.0
Flask==1.1.2
numpy==1.18.1
Pillow==7.0.0
tensorflow-gpu-estimator==2.2.0
Werkzeug==1.0.0
